swap table implementations
