+++ public/app/plugins/panel/table-old/column_options.html
