+++ public/app/plugins/panel/table-old/editor.html
