+++ public/app/plugins/panel/table-old/editor.ts
