+++ public/app/plugins/panel/table-old/module.ts
