+++ public/app/plugins/panel/table-old/renderer.ts
