--- public/app/plugins/panel/table2/types.ts
@@ -1,9 +1,0 @@
-export interface Options {
-  showHeader: boolean;
-  resizable: boolean;
-}
-
-export interface CustomFieldConfig {
-  width: number;
-  displayMode: string;
-}
